% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cranpacks_20211107}
\alias{cranpacks_20211107}
\title{a listing of packages in CRAN on 11/7/2021}
\format{
character()
}
\usage{
cranpacks_20211107
}
\description{
a listing of packages in CRAN on 11/7/2021
}
\note{
For assistance in discriminating Bioconductor packages
}
\keyword{datasets}
