% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deps.R
\name{get_deps}
\alias{get_deps}
\title{use an installer to obtain all dependencies of an OSCA.* book component}
\usage{
get_deps(x, installer = BiocManager::install)
}
\arguments{
\item{x}{path to an OSCA.* package}

\item{installer}{a function link BiocManager::install (default)}
}
\description{
use an installer to obtain all dependencies of an OSCA.* book component
}
\note{
Uses setwd to run in the base folder of the package
}
