% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pbmc.R
\name{make_pbmc}
\alias{make_pbmc}
\title{ensure that sce.pbmc can be created}
\usage{
make_pbmc(curcache = BiocFileCache::BiocFileCache(ask = FALSE))
}
\arguments{
\item{curcache}{defaults to BiocFileCache::BiocFileCache(ask=FALSE)}
}
\description{
ensure that sce.pbmc can be created
}
