% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scpkginfo_3.13}
\alias{scpkginfo_3.13}
\title{a data.frame with hyperlinks for 436 packages used to compute with OSCA book in Bioconductor 3.13}
\format{
data.frame
}
\usage{
scpkginfo_3.13
}
\description{
a data.frame with hyperlinks for 436 packages used to compute with OSCA book in Bioconductor 3.13
}
\examples{
DT::datatable(scpkginfo_3.13, escape=FALSE, pageLength=7)
}
\keyword{datasets}
