% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_packs.R
\name{show_packs}
\alias{show_packs}
\title{view a datatable of packages involved in book}
\usage{
show_packs(vers = "3.13")
}
\arguments{
\item{vers}{character(1) bioconductor version assumed in use}
}
\description{
view a datatable of packages involved in book
}
