% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wr_bioc_expt}
\alias{wr_bioc_expt}
\title{build HTML anchor for Bioconductor experiment package pages}
\usage{
wr_bioc_expt(x, vers = "3.13")
}
\arguments{
\item{x}{character(1) package name}

\item{vers}{character(1) tag for a version, defaults to '3.13'}
}
\description{
build HTML anchor for Bioconductor experiment package pages
}
