% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone.R
\name{clone_osca}
\alias{clone_osca}
\title{use gert to clone all packages}
\usage{
clone_osca(branch = "RELEASE_3_13")
}
\arguments{
\item{branch}{character(1) defaults to RELEASE_3_13}
}
\value{
invisibly the result of `git_clones` in a list
}
\description{
use gert to clone all packages
}
\note{
runs `gert::git_clone` with path set to value of `getwd()`
}
