% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_vignettes.R
\name{run_vignettes_code}
\alias{run_vignettes_code}
\title{run all code in book vignettes in an OSCA.* component}
\usage{
run_vignettes_code(x, exclude = "index.Rmd", purge = TRUE)
}
\arguments{
\item{x}{path to OSCA.* component base folder}

\item{exclude}{character() defaults to `index.Rmd` which would not be compiled in standard usage}

\item{purge}{logical(1) if TRUE, will clean objects produced from global environment and run gc()}
}
\description{
run all code in book vignettes in an OSCA.* component
}
